remove premature pessimization (postinc)
Postinc "i++" is often used when not necessary. In many cases the original
value, before the incrementation is not used. So as a rule of thumb preinc
should be used to avoid the additional overhead of storing, returning
and discarding the original value prior to the incrementation.

See Herb Sutter:
C++ Coding Standards: 101, No. 9. Don’t pessimize prematurely