use _WIN32 instead of WIN32 to activate windows specific features
_WIN32 is defined by the compiler to tell the code that it is compiled under
a 32 bit version of Windows and later. Thus it is also defined under a 64
Bit version of Windows. It's also used by minGW. WIN32 (without underscore)
is *not* defined automatically neither by VC++ nor by minGW. Which is probably
the reason the code compiled in teh first place, because the __stdcall
had been misplaced.

See:
https://stackoverflow.com/questions/9025708/mingw-not-defining-win32-error-in-preprocessor-directives