move the calling convention macro to the correct place
When explicitly giving a calling convention it has to directly precede the
functions name like this:

void __stdcall foo();

The calling convention had been baked into the USB_DMX_DLL macro that
defines the dll import/export specifics resulting to:

__declspec(dllexport) __stdcall void foo();

which in turn leads to a compiler error. Apparently this did not show up
in minGW, because the WIN32 macro had been used to determine the windows
version, not the _WIN32 macro. The WIN32 macro however had not been
defined.

This has been corrected using two dedicated macros, one for the dll magic
and another one for the calling convention.

Additionally the dll magic has been improved to either turn to the
dllexport when compiling the implementation or to dllimport, when using the
header from a third party client. This is chosen by the
USB_DMX_IMPLEMENTATION macro that is defined in the implementation file
before including the usbdmx.h file

A similar error had been in the callback typedefs. Here the calling
convention has to go inside the brackets:

  typedef void (__stdcall foo)(); // correct

instead of

  typedef void __stdcall (foo)(); // compiler error